Rename clerk to informant and implement talk ability
