+++ lib/ruby_warrior.rb
@@ -26,7 +26,7 @@ require 'ruby_warrior/units/thick_sludge'
 require 'ruby_warrior/units/captive'
 require 'ruby_warrior/units/wizard'
 require 'ruby_warrior/units/golem'
-require 'ruby_warrior/units/clerk'
+require 'ruby_warrior/units/informant'
 
 require 'ruby_warrior/abilities/base'
 require 'ruby_warrior/abilities/walk'
@@ -46,5 +46,4 @@ require 'ruby_warrior/abilities/direction_of'
 require 'ruby_warrior/abilities/explode'
 require 'ruby_warrior/abilities/detonate'
 require 'ruby_warrior/abilities/form'
-require 'ruby_warrior/abilities/answer'
 require 'ruby_warrior/abilities/talk'
