--- lib/ruby_warrior/abilities/answer.rb
@@ -1,16 +1,0 @@
-module RubyWarrior
-  module Abilities
-    class Answer < Base
-      attr_accessor :response
-
-      def description
-        "Allows a clerk to share information if talked to."
-      end
-
-      def perform(direction = :forward)
-        #TODO
-        puts response
-      end
-    end
-  end
-end
