+++ towers/clio/level_001.rb
@@ -3,9 +3,9 @@
 # |            |
 # |            |
 # |            |
-# |         C  |
+# |         i  |
 # |           >|
-# |         C  |
+# |         i  |
 # |            |
 # |            |
 # |            |
@@ -22,25 +22,24 @@ description 'As {warrior name} reaches out to take the princess\' hand, a trumpe
 'Finally, the King reaches them. "Don\'t stay with your hand up like that, you look like an idiot"\n'\
 '"When I sent the message out that my daughter needed to be rescued by a brave warrior, I meant my other daughter, Clio"\n'\
 '"She\'s in our vacation castle, or so I heard."\n'\
-'"And by the way, what do you think you did in my castle? Why did you go and murder all our poor guards and clerks? Explain yourself!"\n'\
+'"And by the way, what do you think you did in my castle? Why did you go and murder all our poor guards and informants? Explain yourself!"\n'\
 'After a few seconds of painful silence, the King resumes\n'\
 '"Oh Lord, another one of these mute warriors. I\'ll assume you are just confused, and in warrior fashion you considered anything not wearing a dress as a threat."\n'\
 '"Fine, to redeem yourself you can go ahead and save my Clio. She\'s not here, so there\'s no point in killing anyone else in the castle."\n'\
-'"But on your way out, do me a favor. I need a survey of the kingdom done. See my clerk out there for the details."\n'\
+'"But on your way out, do me a favor. I need a survey of the kingdom done. See my informant out there for the details."\n'\
 'He then leaves the room.\n'
-tip "Use warrior.talk to (maybe) get additional information from clerks"
+tip "Use warrior.talk to (maybe) get additional information from informants"
 
 time_bonus 20 #TODO: Refine
 ace_score 19 #TODO: Refine
-size 12, 12
-stairs 11, 6
+size 4, 1
+stairs 3, 0
 
-warrior 0, 1, :east do |u|
+warrior 0, 0, :east do |u|
   u.add_abilities :walk!, :feel, :direction_of_stairs, :talk!
 end
 
-unit :clerk, 10, 5, :west do |u|
-  u.add_abilities :answer!
-  u.abilities[:answer!].response = "bla bla bla"
+unit :informant, 2, 0, :west do |u|
+  u.response = "blah blah blah"
 end
-unit :clerk, 10, 7, :west
+# unit :informant, 10, 1, :west
